+++ src/main/php/lang/ast/emit/PHP73.class.php
@@ -19,6 +19,7 @@ class PHP73 extends PHP {
     NonCapturingCatchVariables,
     NullsafeAsTernaries,
     OmitArgumentNames,
+    OmitConstantTypes,
     OmitPropertyTypes,
     ReadonlyProperties,
     ReadonlyClasses,
@@ -26,6 +27,7 @@ class PHP73 extends PHP {
     RewriteEnums,
     RewriteExplicitOctals,
     RewriteLambdaExpressions,
+    RewriteStaticVariableInitializations,
     RewriteThrowableExpressions
   ;