+++ src/main/php/lang/ast/emit/PHP74.class.php
@@ -17,12 +17,14 @@ class PHP74 extends PHP {
     NonCapturingCatchVariables,
     NullsafeAsTernaries,
     OmitArgumentNames,
+    OmitConstantTypes,
     ReadonlyClasses,
     ReadonlyProperties,
     RewriteBlockLambdaExpressions,
     RewriteClassOnObjects,
     RewriteEnums,
     RewriteExplicitOctals,
+    RewriteStaticVariableInitializations,
     RewriteThrowableExpressions
   ;