+++ src/main/php/xp/compiler/CompileRunner.class.php
@@ -40,6 +40,9 @@ use util\profiling\Timer;
  *   $ xp compile -t php:7.4 -a php:xp-meta -o dist src/main/php
  *   ```
  *
+ * Use *-e* to change the file extension for generated files inside folders
+ * and archives, which defaults to `.class.php`.
+ *
  * The *-o* and *-n* options accept multiple input sources following them.
  * The *-q* option suppresses all diagnostic output except for errors.
  * 
@@ -69,6 +72,7 @@ class CompileRunner {
     $in= $out= '-';
     $quiet= false;
     $augment= [];
+    $ext= \xp::CLASS_FILE_EXT;
     for ($i= 0; $i < sizeof($args); $i++) {
       if ('-t' === $args[$i]) {
         $target= $args[++$i];
@@ -78,6 +82,8 @@ class CompileRunner {
         $out= $args[++$i];
         $in= array_slice($args, $i + 1);
         break;
+      } else if ('-e' === $args[$i]) {
+        $ext= $args[++$i];
       } else if ('-n' === $args[$i]) {
         $out= null;
         $in= array_slice($args, $i + 1);
@@ -98,7 +104,7 @@ class CompileRunner {
     }
 
     $input= Input::newInstance($in);
-    $output= Output::newInstance($out);
+    $output= Output::newInstance($out)->using($ext);
 
     $t= new Timer();
     $total= $errors= 0;