+++ src/test/php/lang/ast/unittest/emit/EnumTest.class.php
@@ -1,21 +1,22 @@
 <?php namespace lang\ast\unittest\emit;
 
-use lang\reflect\TargetInvocationException;
+use lang\reflection\{Kind, InvocationFailed};
 use lang\{Enum, Error};
-use test\verify\Condition;
-use test\{Action, Assert, Expect, Ignore, Test, Values};
+use test\verify\{Condition, Runtime};
+use test\{Action, Assert, Expect, Test, Values};
 
 #[Condition(assert: 'function_exists("enum_exists")')]
 class EnumTest extends EmittingTest {
+  use AnnotationsOf;
 
   #[Test]
   public function enum_type() {
-    Assert::true($this->type('enum <T> { }')->isEnum());
+    Assert::equals(Kind::$ENUM, $this->declare('enum %T { }')->kind());
   }
 
   #[Test]
   public function name_property() {
-    $t= $this->type('enum <T> {
+    $t= $this->declare('enum %T {
       case Hearts;
       case Diamonds;
       case Clubs;
@@ -26,12 +27,12 @@ class EnumTest extends EmittingTest {
       }
     }');
 
-    Assert::equals('Hearts', $t->getMethod('run')->invoke(null));
+    Assert::equals('Hearts', $t->method('run')->invoke(null));
   }
 
   #[Test]
   public function cases_method_for_unit_enums() {
-    $t= $this->type('enum <T> {
+    $t= $this->declare('enum %T {
       case Hearts;
       case Diamonds;
       case Clubs;
@@ -40,13 +41,13 @@ class EnumTest extends EmittingTest {
 
     Assert::equals(
       ['Hearts', 'Diamonds', 'Clubs', 'Spades'],
-      array_map(function($suit) { return $suit->name; }, $t->getMethod('cases')->invoke(null))
+      array_map(function($suit) { return $suit->name; }, $t->method('cases')->invoke(null))
     );
   }
 
   #[Test]
   public function cases_method_for_backed_enums() {
-    $t= $this->type('enum <T>: string {
+    $t= $this->declare('enum %T: string {
       case Hearts = "♥";
       case Diamonds = "♦";
       case Clubs = "♣";
@@ -55,13 +56,13 @@ class EnumTest extends EmittingTest {
 
     Assert::equals(
       ['Hearts', 'Diamonds', 'Clubs', 'Spades'],
-      array_map(function($suit) { return $suit->name; }, $t->getMethod('cases')->invoke(null))
+      array_map(function($suit) { return $suit->name; }, $t->method('cases')->invoke(null))
     );
   }
 
   #[Test]
   public function cases_method_does_not_yield_constants() {
-    $t= $this->type('enum <T> {
+    $t= $this->declare('enum %T {
       case Hearts;
       case Diamonds;
       case Clubs;
@@ -72,13 +73,13 @@ class EnumTest extends EmittingTest {
 
     Assert::equals(
       ['Hearts', 'Diamonds', 'Clubs', 'Spades'],
-      array_map(function($suit) { return $suit->name; }, $t->getMethod('cases')->invoke(null))
+      array_map(function($suit) { return $suit->name; }, $t->method('cases')->invoke(null))
     );
   }
 
   #[Test]
   public function used_as_parameter_default() {
-    $t= $this->type('enum <T> {
+    $t= $this->declare('enum %T {
       case ASC;
       case DESC;
 
@@ -87,12 +88,12 @@ class EnumTest extends EmittingTest {
       }
     }');
 
-    Assert::equals('ASC', $t->getMethod('run')->invoke(null));
+    Assert::equals('ASC', $t->method('run')->invoke(null));
   }
 
   #[Test]
   public function overwritten_parameter_default_value() {
-    $t= $this->type('enum <T> {
+    $t= $this->declare('enum %T {
       case ASC;
       case DESC;
 
@@ -101,12 +102,12 @@ class EnumTest extends EmittingTest {
       }
     }');
 
-    Assert::equals('DESC', $t->getMethod('run')->invoke(null, [Enum::valueOf($t, 'DESC')]));
+    Assert::equals('DESC', $t->method('run')->invoke(null, [Enum::valueOf($t, 'DESC')]));
   }
 
   #[Test]
   public function value_property_of_backed_enum() {
-    $t= $this->type('enum <T>: string {
+    $t= $this->declare('enum %T: string {
       case ASC  = "asc";
       case DESC = "desc";
 
@@ -115,45 +116,45 @@ class EnumTest extends EmittingTest {
       }
     }');
 
-    Assert::equals('desc', $t->getMethod('run')->invoke(null));
+    Assert::equals('desc', $t->method('run')->invoke(null));
   }
 
   #[Test, Values([[0, 'NO'], [1, 'YES']])]
   public function backed_enum_from_int($arg, $expected) {
-    $t= $this->type('enum <T>: int {
+    $t= $this->declare('enum %T: int {
       case NO  = 0;
       case YES = 1;
     }');
 
-    Assert::equals($expected, $t->getMethod('from')->invoke(null, [$arg])->name);
+    Assert::equals($expected, $t->method('from')->invoke(null, [$arg])->name);
   }
 
   #[Test, Values([['asc', 'ASC'], ['desc', 'DESC']])]
   public function backed_enum_from_string($arg, $expected) {
-    $t= $this->type('enum <T>: string {
+    $t= $this->declare('enum %T: string {
       case ASC  = "asc";
       case DESC = "desc";
     }');
 
-    Assert::equals($expected, $t->getMethod('from')->invoke(null, [$arg])->name);
+    Assert::equals($expected, $t->method('from')->invoke(null, [$arg])->name);
   }
 
   #[Test, Expect(class: Error::class, message: '/"illegal" is not a valid backing value for enum .+/')]
   public function backed_enum_from_nonexistant() {
-    $t= $this->type('enum <T>: string {
+    $t= $this->declare('enum %T: string {
       case ASC  = "asc";
       case DESC = "desc";
     }');
     try {
-      $t->getMethod('from')->invoke(null, ['illegal']);
-    } catch (TargetInvocationException $e) {
+      $t->method('from')->invoke(null, ['illegal']);
+    } catch (InvocationFailed $e) {
       throw $e->getCause();
     }
   }
 
   #[Test, Values([['asc', 'ASC'], ['desc', 'DESC'], ['illegal', null]])]
   public function backed_enum_tryFrom($arg, $expected) {
-    $t= $this->type('enum <T>: string {
+    $t= $this->declare('enum %T: string {
       case ASC  = "asc";
       case DESC = "desc";
 
@@ -162,12 +163,12 @@ class EnumTest extends EmittingTest {
       }
     }');
 
-    Assert::equals($expected, $t->getMethod('run')->invoke(null, [$arg]));
+    Assert::equals($expected, $t->method('run')->invoke(null, [$arg]));
   }
 
   #[Test]
   public function declare_method_on_enum() {
-    $t= $this->type('enum <T> {
+    $t= $this->declare('enum %T {
       case Hearts;
       case Diamonds;
       case Clubs;
@@ -185,36 +186,36 @@ class EnumTest extends EmittingTest {
       }
     }');
 
-    Assert::equals('red', $t->getMethod('run')->invoke(null));
+    Assert::equals('red', $t->method('run')->invoke(null));
   }
 
   #[Test]
   public function enum_implementing_interface() {
-    $t= $this->type('use lang\Closeable; enum <T> implements Closeable {
+    $t= $this->declare('use lang\Closeable; enum %T implements Closeable {
       case File;
       case Stream;
 
       public function close() { }
     }');
 
-    Assert::true($t->isSubclassOf('lang.Closeable'));
+    Assert::true($t->is('lang.Closeable'));
   }
 
   #[Test]
   public function enum_annotations() {
-    $t= $this->type('#[Test] enum <T> { }');
-    Assert::equals(['test' => null], $t->getAnnotations());
+    $t= $this->declare('#[Test] enum %T { }');
+    Assert::equals(['Test' => []], $this->annotations($t));
   }
 
-  #[Test, Ignore('XP core reflection does not support constant annotations')]
+  #[Test, Runtime(php: '>=8.1')]
   public function enum_member_annotations() {
-    $t= $this->type('enum <T> { #[Test] case ONE; }');
-    Assert::equals(['test' => null], $t->getConstant('ONE')->getAnnotations());
+    $t= $this->declare('enum %T { #[Test] case ONE; }');
+    Assert::equals(['Test' => []], $this->annotations($t->constant('ONE')));
   }
 
   #[Test]
   public function cannot_be_cloned() {
-    $t= $this->type('use lang\IllegalStateException; enum <T> {
+    $t= $this->declare('use lang\IllegalStateException; enum %T {
       case ONE;
 
       public static function run() {
@@ -229,13 +230,13 @@ class EnumTest extends EmittingTest {
 
     Assert::equals(
       'Trying to clone an uncloneable object of class '.$t->literal(),
-      $t->getMethod('run')->invoke(null)
+      $t->method('run')->invoke(null)
     );
   }
 
   #[Test]
   public function enum_values() {
-    $t= $this->type('enum <T> {
+    $t= $this->declare('enum %T {
       case Hearts;
       case Diamonds;
       case Clubs;
@@ -250,7 +251,7 @@ class EnumTest extends EmittingTest {
 
   #[Test]
   public function enum_value() {
-    $t= $this->type('enum <T> {
+    $t= $this->declare('enum %T {
       case Hearts;
       case Diamonds;
       case Clubs;