+++ src/test/php/lang/ast/unittest/emit/InstanceOfTest.class.php
@@ -6,7 +6,7 @@ class InstanceOfTest extends EmittingTest {
 
   #[Test]
   public function this_is_instanceof_self() {
-    $r= $this->run('class <T> {
+    $r= $this->run('class %T {
       public function run() {
         return $this instanceof self;
       }
@@ -17,7 +17,7 @@ class InstanceOfTest extends EmittingTest {
 
   #[Test]
   public function new_self_is_instanceof_this() {
-    $r= $this->run('class <T> {
+    $r= $this->run('class %T {
       public function run() {
         return new self() instanceof $this;
       }
@@ -28,7 +28,7 @@ class InstanceOfTest extends EmittingTest {
 
   #[Test]
   public function instanceof_qualified_type() {
-    $r= $this->run('class <T> {
+    $r= $this->run('class %T {
       public function run() {
         return new \util\Date() instanceof \util\Date;
       }
@@ -39,7 +39,7 @@ class InstanceOfTest extends EmittingTest {
 
   #[Test]
   public function instanceof_imported_type() {
-    $r= $this->run('use util\Date; class <T> {
+    $r= $this->run('use util\Date; class %T {
       public function run() {
         return new Date() instanceof Date;
       }
@@ -50,7 +50,7 @@ class InstanceOfTest extends EmittingTest {
 
   #[Test]
   public function instanceof_aliased_type() {
-    $r= $this->run('use util\Date as D; class <T> {
+    $r= $this->run('use util\Date as D; class %T {
       public function run() {
         return new D() instanceof D;
       }
@@ -61,7 +61,7 @@ class InstanceOfTest extends EmittingTest {
 
   #[Test]
   public function instanceof_instance_expr() {
-    $r= $this->run('class <T> {
+    $r= $this->run('class %T {
       private $type= self::class;
 
       public function run() {
@@ -74,7 +74,7 @@ class InstanceOfTest extends EmittingTest {
 
   #[Test]
   public function instanceof_scope_expr() {
-    $r= $this->run('class <T> {
+    $r= $this->run('class %T {
       private static $type= self::class;
 
       public function run() {
@@ -87,7 +87,7 @@ class InstanceOfTest extends EmittingTest {
 
   #[Test]
   public function instanceof_expr() {
-    $r= $this->run('class <T> {
+    $r= $this->run('class %T {
       private function type() { return self::class; }
 
       public function run() {