+++ src/test/php/lang/ast/unittest/emit/NamespacesTest.class.php
@@ -7,17 +7,17 @@ class NamespacesTest extends EmittingTest {
 
   #[Test]
   public function without_namespace() {
-    Assert::equals('', $this->type('class <T> { }')->getPackage()->getName());
+    Assert::null($this->declare('class %T { }')->package());
   }
 
   #[Test]
   public function with_namespace() {
-    Assert::equals('test', $this->type('namespace test; class <T> { }')->getPackage()->getName());
+    Assert::equals('test', $this->declare('namespace test; class %T { }')->package()->name());
   }
 
   #[Test]
   public function resolves_unqualified() {
-    $r= $this->run('namespace util; class <T> {
+    $r= $this->run('namespace util; class %T {
       public function run() {
         return new Date("1977-12-14");
       }
@@ -27,7 +27,7 @@ class NamespacesTest extends EmittingTest {
 
   #[Test]
   public function resolves_relative() {
-    $r= $this->run('class <T> {
+    $r= $this->run('class %T {
       public function run() {
         return new util\Date("1977-12-14");
       }
@@ -37,7 +37,7 @@ class NamespacesTest extends EmittingTest {
 
   #[Test]
   public function resolves_absolute() {
-    $r= $this->run('namespace test; class <T> {
+    $r= $this->run('namespace test; class %T {
       public function run() {
         return new \util\Date("1977-12-14");
       }
@@ -47,7 +47,7 @@ class NamespacesTest extends EmittingTest {
 
   #[Test]
   public function resolves_import() {
-    $r= $this->run('namespace test; use util\Date; class <T> {
+    $r= $this->run('namespace test; use util\Date; class %T {
       public function run() {
         return new Date("1977-12-14");
       }
@@ -57,7 +57,7 @@ class NamespacesTest extends EmittingTest {
 
   #[Test]
   public function resolves_alias() {
-    $r= $this->run('namespace test; use util\Date as DateTime; class <T> {
+    $r= $this->run('namespace test; use util\Date as DateTime; class %T {
       public function run() {
         return new DateTime("1977-12-14");
       }
@@ -67,7 +67,7 @@ class NamespacesTest extends EmittingTest {
 
   #[Test]
   public function resolves_namespace_keyword() {
-    $r= $this->run('namespace util; class <T> {
+    $r= $this->run('namespace util; class %T {
       public function run() {
         return new namespace\Date("1977-12-14");
       }
@@ -77,7 +77,7 @@ class NamespacesTest extends EmittingTest {
 
   #[Test]
   public function resolves_sub_namespace() {
-    $r= $this->run('class <T> {
+    $r= $this->run('class %T {
       public function run() {
         return new namespace\util\Date("1977-12-14");
       }