+++ src/test/php/lang/ast/unittest/emit/ReturnTest.class.php
@@ -6,7 +6,7 @@ class ReturnTest extends EmittingTest {
 
   #[Test]
   public function return_literal() {
-    $r= $this->run('class <T> {
+    $r= $this->run('class %T {
       public function run() {
         return "Test";
       }
@@ -16,7 +16,7 @@ class ReturnTest extends EmittingTest {
 
   #[Test]
   public function return_member() {
-    $r= $this->run('class <T> {
+    $r= $this->run('class %T {
       private $member= "Test";
 
       public function run() {
@@ -28,7 +28,7 @@ class ReturnTest extends EmittingTest {
 
   #[Test]
   public function return_without_expression() {
-    $r= $this->run('class <T> {
+    $r= $this->run('class %T {
       public function run() {
         return;
       }